feat(gcp): create esc environment with oidc details
