+++ meta/Pulumi.dev.yaml
@@ -1,6 +1,3 @@
 ---
 environment:
   - pulumi-support
-config:
-  azure-native:location: WestUS2
-  meta:environmentName: pulumi-support
