+++ meta/README.md
@@ -1,8 +1,27 @@
-# GCP configuration
+# Pulumi-home meta
 
-Creates a small project within the pulumi org.
+Configures a working environment for pulumi support tasks
 
-Enables the Google Drive service.
+- [gcp/auth](###Auth)
+- [gcp/projects](###Projects)
+- [gcp/services](###Services)
+
+## GCP
+
+#### Auth
+
+Creates OIDC configuration for authenticating with GCP using pulumi as an
+identity provider. Creates a Pulumi ESC environment to integrate the OIDC into
+other pulumi projects.
+
+#### Projects
+
+Creates an gcp project within the pulumi-ce organization to avoid breaking other
+ppeople's stuff.
+
+#### Services
+
+Enables the gdrive_api_service on the project created in [Projects](#projects).
 
 Currently you need to make the OAuth credentials by hand on the
 [Credentials](https://console.cloud.google.com/apis/credentials?project=pulumi-ce-stoo)
